PhoneApp: Disable the use of Proximity Sensor in deciding InCall Screen time out.
Screen timeout value needs to be based on the InCall time out values and not
on the availablility of the Proximity sensor during development.
If the value is decided based on the input from the Proximity sensor then the
Screen will not time out at the expected time out period of 15 seconds.

Change-Id: Ida54e7a8b1d057af4737dbaaadf77790432ccb15