Fix hard-to-reproduce crash with emergency call in airplane mode
If you make an emergency call while in airplane mode, the phone app uses a
helper activity called EmergencyCallHandler that knows how to force you
out of airplane mode and then retry the call.

When launching the EmergencyCallHandler, we pass it an intent that it
should use later when re-launching the call.  That happens in two places:
(1) in the placeCall() method, before even trying to place a call, if we
notice the radio is off, and (2) in the onDisconnect() method if we did
try the call but it failed with the disconnect reason OUT_OF_SERVICE.

The intent we use is basically the intent we were originally launched
with.  Unfortunately, the code in case (2) was calling getIntent() to look
up that intent, but by that point we had already reset the "current
intent" to its initial UNDEFINED state :-(

The fix is to stash away the InCallScreen's original intent at the top of
onDisconnect(), and use that intent when launching the EmergencyCallHandler.

This bug is very rare in practice, by the way (and till recently was only
ever seen on early crespo prototypes) because case (1) is the most common
code path.  Case (2) only runs if we re-try the call after turning on the
radio, but it still fails, and we have to try again.

Note the actual fix here is pretty small: just the lines of code in
InCallScreen.onDisconnect() where we use "originaInCallIntent" instead of
getIntent() when launching the EmergencyCallHandler.  Everything else in
this change is comment updates and debug logging cleanup.

TESTED:
  - normal 911 call: dialed successfully.
  - dialing 911 while in airplane mode: turned radio on first, then dialed
    successfully.
(Note I ran these tests using a hacked version of GSMPhone.java that
redirected 911 calls to a harmless regular number.)

Bug: 3094858
Change-Id: Ib2f7168d943cc8eeddb30455c85dfd868f634a23